Fix: Clicking the modify button on the Recipe tab will load the selected recipe into the RecipeForm.
At this time the Ingredient rows will populate based off the information in the RecipeIngredient Table with -!!- one key note -!!- ; the Unit combo box will not correctly display the unit metric associated with the Ingredient upon the form's load. I'm not sure as to the cause, but I think it might be related to some of the backend code. Once I get the RecipeForm to actually create new Inventory records for items that aren't recorded in the Inventory table, this issue may resolve itself (or it could be some obscure WinForm issue which requires a very specific and awkward way of setting its default value).

Other note: I could've just been trying to solve it for too long and am totally missing the obvious solution.